Render input value with PhoneNumberPrefixWidget on form errors
The phone number input field was incorrectly cleared when:
- a form has an error on another field, and
- the phone number field is displayed with PhoneNumberPrefixWidget, and
- the phone number region is the default region.

That behavior was caused by 005769cf39323e5b23710783f45befb546672cd6,
which incorrectly used `prepare_value()` to transform a PhoneNumber
instance to an str (the phone number in national format). By doing so,
information was lost and the decompress() method of
PhoneNumberPrefixWidget unexpectedly received an str.

The phone number value is now always a PhoneNumber instance. It’s up to
the widget to format that value.
To retain the behavior of showing national phone numbers in the national
format by default, the formfields.PhoneNumberField widget is now
PhoneNumberInternationalFallbackWidget.

PhoneNumberInternationalFallbackWidget behavior was updated to match the
previous implementation of prepare_value: the number is displayed in
national format if its country code matches the country code of the
configured region.
It now represents its value as a PhoneNumber, a richer type than `str`,
which facilitates the custom display logic.